// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface LastMinutePackage {
  id: number;
  destination?: string | null;
  package?: string | null;
  discount?: number | null;
  order?: number | null;
  features?:
    | {
        feature?: string | null;
        id?: string | null;
      }[]
    | null;
  cost?: number | null;
  image?: (number | null) | Media;
  gallery?: (number | Media)[] | null;
  overview?: string | null;
  highlights?: {
    chips?:
      | {
          chip?: string | null;
          id?: string | null;
        }[]
      | null;
    pointers?:
      | {
          pointer?: string | null;
          id?: string | null;
        }[]
      | null;
  };
  day_breakup?: {
    days?:
      | {
          title?: string | null;
          break_up?:
            | {
                title?: string | null;
                subtitle?: string | null;
                id?: string | null;
              }[]
            | null;
          id?: string | null;
        }[]
      | null;
  };
  accomodations?: {
    locations?:
      | {
          image?: (number | null) | Media;
          name?: string | null;
          location?: string | null;
          type?: string | null;
          whats_included?: string | null;
          link?: string | null;
          id?: string | null;
        }[]
      | null;
  };
  updatedAt: string;
  createdAt: string;
}